refactor: rename v1.4 bundles types
