+++ app/upgrades/v1_5/v1_4_bundles_types/getters_params.go
@@ -1,4 +1,4 @@
-package v1_4_types
+package v1_4_bundles_types
 
 import (
 	storeTypes "cosmossdk.io/store/types"
